refactor(shared): implement phase 11 code quality standards
- Create type-only enums.ts with Role and LegislativeSource type aliases
- Inline Role, LegislativeSource, REVALIDATE constants in index.ts (webpack-safe)
- Update all shared type files to import and use Role/LegislativeSource types
- Remove constants.ts (inlined into index.ts for webpack compatibility)
- Add transpilePackages config to next.config.ts

Implementation fulfills all acceptance criteria:
✓ Role enum with DEPUTADO='deputado', SENADOR='senador'
✓ LegislativeSource enum with CAMARA='camara', SENADO='senado'
✓ Zero raw string unions in type files
✓ All 5 cursor decodeCursor functions use Zod schema.parse()
✓ source.service uses SourceStatusValueSchema.parse()
✓ REVALIDATE constants exported and used in api-client
✓ All 131 tests pass
✓ vercel build passes
✓ Cursor validation tests verify malformed cursors rejected

This is a non-functional refactor: no new features, all tests remain green.
